requirements.txt
discord.py==2.3.2
python-dotenv==1.0.0
PyNaCl==1.5.0
certifi==2024.2.2
wavelink==2.6.3
